feat: Implementa CurrentUserService e lógica de acesso aos métodos de usuário

+++ src/Application/Services/UserService.cs
@@ -2,57 +2,88 @@
 using CashFlowAPI.Application.Dtos;
 using CashFlowAPI.Application.Interfaces;
 using CashFlowAPI.Domain.Entities;
+using CashFlowAPI.Domain.Enum;
 using CashFlowAPI.Domain.Interfaces;
-using CashFlowAPI.Domain.Security;
 
 namespace CashFlowAPI.Application.Services;
 public sealed class UserService : IUserService
 {
     private readonly IUserRepository _userRepository;
     private readonly IMapper _mapper;
+    private readonly ICurrentUserService _currentUserService;
 
-    public UserService(IUserRepository userRepository, IMapper mapper)
+    public UserService(IUserRepository userRepository, IMapper mapper, ICurrentUserService currentUserService)
     {
+        _currentUserService = currentUserService;
         _userRepository = userRepository;
         _mapper = mapper;
     }
 
     public async Task<UserDto> CreateUserAsync(UserDto userDto)
     {
-        if (await _userRepository.AnyAsync(x => x.Name == userDto.Name))
-            throw new Exception("O usuário já está cadastrado.");
-        var user = _mapper.Map<User>(userDto);
-        user.Validate();
-        user.SetPassword(userDto.Password.Value);
-        await _userRepository.AddAsync(user);
-        return _mapper.Map<UserDto>(user);
+        var currentUser = await _currentUserService.GetCurrentUser();
+        if (userDto.Role == UserRoles.Admin && currentUser == null || userDto.Role == UserRoles.Admin && currentUser.Role != "Admin")
+            throw new UnauthorizedAccessException("Usuário não tem permissão para criar um novo administrador.");
+        else
+        {
+            if (await _userRepository.AnyAsync(x => x.Email.Value == userDto.Email.Value))
+                throw new Exception("O email de usuário já está cadastrado.");
+
+            var user = _mapper.Map<User>(userDto);
+            user.Validate();
+            user.SetPassword(userDto.Password.Value);
+            await _userRepository.AddAsync(user);
+            return _mapper.Map<UserDto>(user);
+        }
     }
+
     public async Task<UserDto> UpdateUserAsync(UserDto userDto, Guid userId)
     {
-        var user = await _userRepository.GetByIdAsync(userId);
-        user.Update(userDto.Name, userDto.Email, userDto.Password);
-        user.Validate();
-        user.SetPassword(userDto.Password.Value);
-        if (user == null)
-            throw new Exception("O usuario não foi encontrado.");
-        await _userRepository.UpdateAsync(user);
-        return _mapper.Map<UserDto>(user);
+        var currentUser = await _currentUserService.GetCurrentUser();
+        if (userId != currentUser.UserId && currentUser.Role != "Admin")
+            throw new UnauthorizedAccessException("Somente administradores podem alterar outros usuários.");
+        else
+        {
+            var user = await _userRepository.GetByIdAsync(userId);
+            user.Update(userDto.Name, userDto.Email, userDto.Password);
+            user.Validate();
+            user.SetPassword(userDto.Password.Value);
+            if (user == null)
+                throw new Exception("O usuario não foi encontrado.");
+            await _userRepository.UpdateAsync(user);
+            return _mapper.Map<UserDto>(user);
+        }
     }
+
     public async Task<UserDto> DeleteUserAsync(Guid userId)
     {
-        var user = await _userRepository.GetByIdAsync(userId);
-        if (user == null)
-            throw new Exception("O usuario não foi encontrado.");
-        await _userRepository.DeleteAsync(userId);
-        return null;
+        var currentUser = await _currentUserService.GetCurrentUser();
+        if (userId != currentUser.UserId && currentUser.Role != "Admin")
+            throw new UnauthorizedAccessException("Somente administradores podem deletar outros usuários.");
+        else
+        {
+            var user = await _userRepository.GetByIdAsync(userId);
+            if (user == null)
+                throw new Exception("O usuario não foi encontrado.");
+            await _userRepository.DeleteAsync(userId);
+            return null;
+        }
     }
+
     public async Task<UserDto> GetUserById(Guid userId)
     {
-        var user = await _userRepository.GetUserById(userId);
-        if (user == null)
-            throw new Exception("O usuario não foi encontrado.");
-        return _mapper.Map<UserDto>(user);
+        var currentUser = await _currentUserService.GetCurrentUser();
+        if (userId != currentUser.UserId && currentUser.Role != "Admin")
+            throw new UnauthorizedAccessException("Somente administradores podem obter informações de outros usuários.");
+        else
+        {
+            var user = await _userRepository.GetUserById(userId);
+            if (user == null)
+                throw new Exception("O usuario não foi encontrado.");
+            return _mapper.Map<UserDto>(user);
+        }
     }
+
     public async Task<List<UserDto>> GetAllAsync()
     {
         var users = await _userRepository.GetAllUsers();
